refactor(gateway): extract Config loader methods (GW-2 S7)
Move impl Config { keycloak_backend_url, load, validate } to
config/loader.rs along with its regression + load + validate tests.
Cleans figment/path/tracing::info imports from config.rs.
